test: generalize generateTestCases util
