+++ test/schemas/core/address.js
@@ -1,29 +1,7 @@
 'use strict';
 
-const expect = require('chai').expect;
 const schema = require('../../../prebuilt/core/components/address.json');
-const validator = require('../../../validator');
-const ValidationError = require('../../../ValidationError');
-
-function generateTestCases(schema, positive, cases) {
-  if (positive) {
-    it('positive cases', () => {
-      cases.forEach(value => {
-        expect(validator.validateSync(schema, value)).to.be.an.object;
-      });
-    });
-  } else {
-    it('negative cases', () => {
-      cases.forEach(value => {
-        const func = () => {
-          validator.validateSync(schema, value);
-          console.warn(`We should not get here. Input: ${value}`);
-        };
-        expect(func).to.throw(ValidationError);
-      });
-    });
-  }
-}
+const { generateTestCases } = require('../../../test-lib');
 
 describe('address.placeName', () => {
   generateTestCases(schema.definitions.placeName, true, [
